// src/app/modules/banks/models/bank.model.ts
// Generated by https://quicktype.io

export interface Content<T> {
  content:          T[];
  empty:            boolean;
  first:            boolean;
  last:             boolean;
  number:           number;
  numberOfElements: number;
  pageable:         Pageable;
  size:             number;
  sort:             Sort;
  totalElements:    number;
  totalPages:       number;
}

export interface IBank {
  codigo:    string;
  descricao: string;
  id:        number;
  status:    Status;
  isSelected?:    boolean;
}

export interface Status {
  descricao: string;
  id:        string;
}

export interface Pageable {
  offset:     number;
  pageNumber: number;
  pageSize:   number;
  paged:      boolean;
  sort:       Sort;
  unpaged:    boolean;
}

export interface Sort {
  empty:    boolean;
  sorted:   boolean;
  unsorted: boolean;
}